requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.17
twilio==8.10.3
elevenlabs==0.2.26
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.3
PyPDF2==3.0.1
python-docx==1.1.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1